Add PreviousCPUUtilizationPercentage field to HPA
This addition makes it so now `HorizontalPodAutoscalerStatus` contains a
field for `PreviousCPUUtilizationPercentage` and fills it with the
previous value of `CurrentCPUUtilizationPercentage` each time
`CurrentCPUUtilizationPercentage` is updated. This is a simple initial
way of determining the derivative of CPU utilization for in order to
project future CPU utilization when using predictive auto-scaling.

Signed-off-by: mattjmcnaughton <mattjmcnaughton@gmail.com>