[WIP] Remove rkt specific code from pleg
Remove code from the `pleg` which is only necessary when pods don't
contain sandboxes. Both Docker and cri, the only supported runtimes,
use sandboxes, so this code is no longer necessary.

Improve the test coverage to test actually falling back to the
sandboxes, instead of just getting the IP from the pod statuses.